+++ yajl-1.0.11/yajl_buf.h
@@ -30,6 +30,8 @@
  * POSSIBILITY OF SUCH DAMAGE.
  */ 
 
+#include <stdlib.h>
+
 #ifndef __YAJL_BUF_H__
 #define __YAJL_BUF_H__
 
@@ -56,7 +58,7 @@ yajl_buf yajl_buf_alloc(yajl_alloc_funcs * alloc);
 void yajl_buf_free(yajl_buf buf);
 
 /* append a number of bytes to the buffer */
-void yajl_buf_append(yajl_buf buf, const void * data, unsigned int len);
+void yajl_buf_append(yajl_buf buf, const void * data, size_t len);
 
 /* empty the buffer */
 void yajl_buf_clear(yajl_buf buf);
