+++ yajl-1.0.11/yajl_encode.h
@@ -39,7 +39,7 @@
 void yajl_string_encode2(const yajl_print_t printer,
                          void * ctx,
                          const unsigned char * str,
-                         unsigned int length);
+                         size_t length);
 
 void yajl_string_encode(yajl_buf buf, const unsigned char * str,
                         unsigned int length);
